Minor code correction
