+++ components/__pycache__/ai.cpython-310.pyc
