+++ components/__pycache__/assets.cpython-311.pyc
