+++ components/assets.py
